refactor:  moved card component refrence to higher component
