+++ src/app/pages/quickMaps/pages/costEffectiveness/components/interventionComparison/interventionComparison.component.html
@@ -29,18 +29,3 @@
         </button>
     </span>
 </div>
-
-<mat-fab-menu #fabMenu matTooltip="Dashboard Layout" [matTooltipDisabled]="fabMenu.isActive" class="fab-layout"
-    color="accent" [fabButtons]="layoutOptions" icon="dashboard_customize"
-    (onFabMenuItemSelected)="layoutChange($event)">
-</mat-fab-menu>
-<gridster class="bmgf-maps-nice-scrollbar" [options]="options">
-    <gridster-item [item]="widget" *ngFor="let widget of dashboard">
-        <app-card #card [widget]="widget" [resizeEvent]="resizeEvent" [ngSwitch]="widget.type">
-            <!-- Comparison card -->
-            <!-- <app-dc-comparison-card *ngSwitchCase="WIDGETS.DIET_CHANGE_COMPARISON" [card]="card"></app-dc-comparison-card> -->
-            <app-intervention-comparison-card *ngSwitchCase="WIDGETS.COST_EFFECTIVENESS_COMPARISION" [card]="card">
-            </app-intervention-comparison-card>
-        </app-card>
-    </gridster-item>
-</gridster>
