+++ src/app/pages/quickMaps/pages/costEffectiveness/costEffectiveness.component.html
@@ -1,4 +1,16 @@
-<div class="bmgf-maps-nice-scrollbar">
+<!-- <mat-fab-menu #fabMenu matTooltip="Dashboard Layout" [matTooltipDisabled]="fabMenu.isActive" class="fab-layout"
+    color="accent" [fabButtons]="layoutOptions" icon="dashboard_customize"
+    (onFabMenuItemSelected)="layoutChange($event)">
+</mat-fab-menu> -->
+<gridster class="bmgf-maps-nice-scrollbar" [options]="options">
     <app-ce-description></app-ce-description>
     <app-ce-menu></app-ce-menu>
-</div>
+    <gridster-item [item]="widget" *ngFor="let widget of dashboard">
+        <app-card #card [widget]="widget" [resizeEvent]="resizeEvent" [ngSwitch]="widget.type">
+            <!-- Comparison card -->
+            <!-- <app-dc-comparison-card *ngSwitchCase="WIDGETS.DIET_CHANGE_COMPARISON" [card]="card"></app-dc-comparison-card> -->
+            <app-intervention-comparison-card *ngSwitchCase="WIDGETS.COST_EFFECTIVENESS_COMPARISION" [card]="card">
+            </app-intervention-comparison-card>
+        </app-card>
+    </gridster-item>
+</gridster>
