+++ src/app/pages/quickMaps/pages/costEffectiveness/costEffectiveness.component.scss
@@ -1,5 +1,5 @@
 :host {
-  padding: 10px;
-  height: 100%;
-  width: 100%;
+  // padding: 10px;
+  // height: 100%;
+  // width: 100%;
 }
